Add hook to restore authorized_keys permissions
Git only stores the execute bit, but the authorized_keys file has to
have permissions of 600 for SSH to function. This hook resets
authorized_keys to 600 if it was changed in the last merge.

I also added a new .githooks directory which allows me to manage the
hooks with Git. New hooks can be symlinked into the real .git/hooks
directory with .githooks/link-hooks.sh