Fix reattach-to-user-namespace in Tmux
Due to a race condition in Tmux, it's not possible to set a
default-command inside an if-shell, so the OS X specific options can't
be loaded dynamically when starting Tmux.

To workaround this, .bashrc sets an alias to first load .tmux.osx.conf
only on OS X. This file sets the default-command (and other OS X
specific options) and sources the main configuration file.