Enable Powerline for both Vim and Tmux
For everything to work, Powerline must also be installed with:
    pip install git+https://github.com/Lokaltog/powerline

Or if pip isn't available, the it can be installed manually with:
    git clone https://github.com/Lokaltog/powerline.git
    cd powerline
    python setup.py build
    python setup.py install

To have this work on Gvim in Windows, the Windows version of Python (not
just Cygwin) needs to be installed, and then the above commands run in
the Windows Python environment.

Also, this requires a Powerline-patched font to look best.
I've made my own Powerline-patched versions of Ricty.